Refactored JS functions, added PHP function and finalized style
Broke positioning function into smaller functions and moved the array (which
represents the data that will eventually be retrieved using MySQL) to a function
to keep index.php clean. Added code to resize the submenu containing div
depending on the width of the inside elements, rather than having the width hard
coded in the CSS. Settled on dark/neon styling to call to mind a console.